+++ lib/github/create-policies-pr.ts
@@ -1,8 +1,7 @@
 import { serverInstance as rollbar } from '@/lib/rollbar';
 import { Octokit } from '@octokit/rest';
 
-import { PlatformPolicy } from '@/lib/platform-policies/types';
-import { ValidationSession, PolicyUpdate } from '@/lib/ai/policy-validation';
+import { PlatformPolicy, PolicyUpdate, ValidationSession } from '@/types/policies';
 
 export interface PullRequestData {
   title: string;
@@ -171,7 +170,11 @@ export class GitHubPRCreator {
 
     try {
       const branchName = GitHubPRCreator.generateBranchName(session.validationId);
-      const prTitle = GitHubPRCreator.generatePRTitle(totalChanges, totalPlatformsUpdated, Object.keys(updatedPolicies));
+      const prTitle = GitHubPRCreator.generatePRTitle(
+        totalChanges,
+        totalPlatformsUpdated,
+        Object.keys(updatedPolicies),
+      );
       const prBody = GitHubPRCreator.generatePRBody(
         `Policy validation completed with ${totalChanges} policy changes across ${totalPlatformsUpdated} platforms`,
         session.validationId,
@@ -275,10 +278,15 @@ This PR was automatically generated by the Policy Validation Workflow. The chang
   /**
    * Generates PR title
    */
-  static generatePRTitle(totalChanges: number, totalPlatformsUpdated: number, platformsUpdated: string[]): string {
-    const platformsText = totalPlatformsUpdated === 1
-      ? platformsUpdated[0]?.toUpperCase() || 'platform'
-      : `${totalPlatformsUpdated} platforms`;
+  static generatePRTitle(
+    totalChanges: number,
+    totalPlatformsUpdated: number,
+    platformsUpdated: string[],
+  ): string {
+    const platformsText =
+      totalPlatformsUpdated === 1
+        ? platformsUpdated[0]?.toUpperCase() || 'platform'
+        : `${totalPlatformsUpdated} platforms`;
 
     return `🤖 Policy Update: ${totalChanges} changes across ${platformsText}`;
   }
@@ -287,7 +295,10 @@ This PR was automatically generated by the Policy Validation Workflow. The chang
 /**
  * Applies policy updates to a platform policy object by mapping references
  */
-export function applyPolicyUpdates(platformPolicy: PlatformPolicy, updatedPolicies: PolicyUpdate[]): boolean {
+export function applyPolicyUpdates(
+  platformPolicy: PlatformPolicy,
+  updatedPolicies: PolicyUpdate[],
+): boolean {
   let hasChanges = false;
   const timestamp = new Date().toISOString();
 
@@ -297,7 +308,7 @@ export function applyPolicyUpdates(platformPolicy: PlatformPolicy, updatedPolici
 
   updatedPolicies.forEach((updatedPolicy: any) => {
     const policyRef = updatedPolicy.reference;
-    
+
     platformPolicy.contentTypes?.forEach((contentType) => {
       const policyIndex = contentType.policies?.findIndex((p) => p.reference === policyRef);
       if (policyIndex !== -1) {
@@ -324,7 +335,9 @@ export function applyPolicyUpdates(platformPolicy: PlatformPolicy, updatedPolici
       }
     });
 
-    const generalPolicyIndex = platformPolicy.generalPolicies?.findIndex((p) => p.reference === policyRef);
+    const generalPolicyIndex = platformPolicy.generalPolicies?.findIndex(
+      (p) => p.reference === policyRef,
+    );
     if (generalPolicyIndex !== -1) {
       platformPolicy.generalPolicies[generalPolicyIndex] = {
         ...platformPolicy.generalPolicies[generalPolicyIndex],
